refactor(sync): move sync-v1 files to a separate module
