+++ hathor/p2p/manager.py
@@ -98,8 +98,8 @@ class ConnectionsManager:
                  enable_sync_v1: bool,
                  enable_sync_v2: bool,
                  enable_sync_v1_1: bool) -> None:
-        from hathor.p2p.sync_v1_1_factory import SyncV11Factory
-        from hathor.p2p.sync_v1_factory import SyncV1Factory
+        from hathor.p2p.sync_v1.factory_v1_0 import SyncV10Factory
+        from hathor.p2p.sync_v1.factory_v1_1 import SyncV11Factory
 
         if not (enable_sync_v1 or enable_sync_v1_1 or enable_sync_v2):
             raise TypeError(f'{type(self).__name__}() at least one sync version is required')
@@ -181,11 +181,11 @@ class ConnectionsManager:
         # sync-manager factories
         self._sync_factories = {}
         if enable_sync_v1:
-            self._sync_factories[SyncVersion.V1] = SyncV1Factory(self)
+            self._sync_factories[SyncVersion.V1] = SyncV10Factory(self)
         if enable_sync_v1_1:
             self._sync_factories[SyncVersion.V1_1] = SyncV11Factory(self)
         if enable_sync_v2:
-            self._sync_factories[SyncVersion.V2] = SyncV1Factory(self)
+            self._sync_factories[SyncVersion.V2] = SyncV10Factory(self)
 
     def set_manager(self, manager: 'HathorManager') -> None:
         """Set the manager. This method must be called before start()."""
